Fix API Error CloudWatch alarms by specifying API name correctly (#103)
<!-- Please complete the following sections as necessary. -->

### Description
Fixes the 4XXError and 5XXError alarms by specifying the ApiName
correctly in the alarm config.

### Steps to Test
1. Deploy the changes to `Innov-Platform-Dev`
2. Send invalid requests to the Feedback API (i.e. send a request to the
`/rating` endpoint with an empty request body).
3. Wait a few minutes, then open the CloudWatch alarm in the AWS console
and check that errors appear on the graph.

Note: I was only able to verify the 5XXError alarm since the Feedback
API currently only returns 500 or 400 responses. But since the alarms
are configured identically besides the metric, I believe this is
sufficient to ensure that the 4XXError alarm would work as well.