Merge main into dev in order to clean up commit history (#105)
Not sure why, but GitHub says `main` is 23 commits ahead `dev` yet there
is no diff (files changed are 0). I believe these are an artifact of
previous merges/reverts that occurred on `main` but not on `dev`. So, in
theory, merging this PR only affects commit history and should not
result in any code changes

---------

Co-authored-by: sanni <43077388+sannidhishukla@users.noreply.github.com>
Co-authored-by: AnJu Hyppolite <anjuhyppolite@pursuit.org>
Co-authored-by: AnJu Hyppolite <75342585+AnJuHyppolite@users.noreply.github.com>