Add AWS authentication section to README (#109)
Splits out a "How to authenticate to AWS accounts" section in the
README, which includes more detailed instructions on how to authenticate
using AWS SSO, including a newly created Bitwarden item with a sample
`~/.aws/config`.

Also adds a couple NPM commands for deploying to a particular account
when using AWS SSO to authenticate.